Support CentOS7
Recently, I updated koviz to get rid of warnings
I was getting on Redhat 8.  A lot of the warnings
were due to deprecated usage. Unfortunately,
the move broke CI for some and broke TS21 which
is stuck in CentOS7.

This branch was made to support both
Redhat8 and CentOS7 and keep the compile
clean of warnings/errors on both platforms.