Curve summations started
Dan Fenn hoped for curve summations.  This is a
start, but most working.

When one presses 's', the current set of curves
are added together and added to the plot.

If the curve timestamps do not match, koviz will
interpolate values but will not extrapolate.  The
reason is from Dan's comment:

    My thought is that interpolation is probably acceptable here,
    but extrapolation isn't. Each curve is defined on its domain,
    and once you try to extrapolate outside that domain, you're really
    in undefined territory. So I'd say the sum should only be defined
    on the intersection of the domains

Koviz will also handle duplicate timestamps.  Koviz throws
out NaNs and will not attempt to interpolate them.  The time
alignment honors the "time match tolerance" i.e. the -tmt option.
Koviz will also convert units if it can so adding a curve with
meters and another curve with centimeters works as expected.

The thing left to do is accounting for curve scale/bias.
Also need to make sure the sum works with shifted runs.