Flatline label can be drawn outside viewable window
When doing error+compare plots the Flatline label
was being drawn on top of the curve in any case.
In the case of an error+compare, the flatline could be
at the very top of the window, which pushed the
label offscreen.  This fix simply draws the
Flatline label below the curve if drawing above
goes offscreen.