Monte carlo -beginRun and -endRun options not working
The -beginRun and -endRun commandline options were
not even being used. It looks like one of the big
rewrites failed to take the begin/end run options
into account.  The code was there to handle the
options, I just needed to pass the options to the
class that uses the beg/end run values.