Stop video window from moving on reload
Tim ran into this.  When opening video after
video, the video window would creep/move down
the screen.

The problem was that vidView->geometry() gets the
location of the widget's rectangle, not the
window's geometry.  When vidView->move() occurs,
it's in Window coord space not widget space.
I needed to get the window geometry saved
into lastVideoRect.