Curve combinations hit OS limit when over 1000 curves
The fix for this was simple.  I thought when mem mapping
trk files that one had to keep the file open during the
map, and then close the file when the munmap occured.
When doing curve combinations of 1000+ curves this caused
the OS to error out due to reaching the max number
of open files.  It turns out that one can immediately
close the trk file after the initial mmap call instead
of waiting to close with the munmap call.  This fixed
opening up too many files at once.  The limit now is
how many mem maps one can have, but that limit is
very high, see /proc/sys/vm/max_map_count.  It is
65,530 on most systems.