Change curve combo algorithm and add MC envelope
This is what I wrote Dan:

I'm modifying the behavior of the curve combinator.  The first
cut only combined points if interpolation possible (no extrapolation).
Instead, I'll combine the curves but just not include y values
from points that would have to be extrapolated - but still give a point.
For example, if there are three curves A,B,C with domains [0-6], [3-7], [5,12]
- the resulting combo curve would have domain [0-12].  The points [0-3)
would just have the values from the A curve and (7-12] from the C curve.
That works much better for curves without the same domain - and especially
for generating a upper/lower bound curves from a monte carlo set of curves.