Make showing video optional
When one has a monte carlo set of RUNs
with videos per run, it is desirable to
disable video if one wants to simply
peruse the data without the overhead
of video.

This commit does a few things:
1. Adds a -showVideo commandline option
2. If the -showVideo option is not specified
    1. If -video or -videoList is on the
       commandline, vidoes will be shown
    2. Else koviz will try to use the
       last user (un)checked Options > ShowVideo
       value that is saved in koviz.conf settings
    3. Lastly if the settings file is not present
       koviz will show the video.  If the user
       wants to close the video sync, the user
       can uncheck Options > ShowVideo

While running koviz with video, the user may
always (un)check Options > ShowVideo.  Koviz
will save that choice and use it on subsequent
sessions.