Add DP Tab_Label spec to page and table widget tabs
Hung asked for this.  When one has a DP spec file with
multiple pages, one can label the widget tabs in the UI
to identify each page.  Prior to this enhancement, the
page tab labels just had the DP_name as the tab label.
If DP_name had multiple pages, all the pages had the
same name - making it difficult to navigate.