WinKoviz rendered offscreen if external monitor taken off
Got this report from ARGOS.  If one has koviz on an external
monitor hooked to a laptop, then closes koviz, then unplugs
the monitor.  When koviz is brought back up, koviz tries to
render to the unplugged monitor since it always tries to
place the window where it last was.

To get around this, need to see if the window placement
is viewable before placing.  On Gnome, this isn't a problem
since the Window manager will ignore placing a window
in a non-existent location.