Custom video directory path
Here is how it is documented:

By default, koviz will search for a “video” directory in the
RUN directory. If one would rather place the videos associated
 with the RUN data in another location, one can give koviz a
relative path from the RUN to the video directory. There are
two ways to set the video directory. First, one can use the
commandline option -videoDir. Second, one can set the video
directory path in the main UI window: Options > VideoDirectory.
For example, if the data is structured like below, one would set
the Option > VideoDirectory to “../Video” since the relative
path from the directory with the sim data to the
Video directory is “../Video”:
    RUN_a/
        DATA/
            log_Mars_environment.trk
            log_Mars_rover.trk
        Video/
            video-offsets.txt
            video-1.mp4
            video-2.mp4