More marker unit tests and tweaks
The biggest change for this commit was
simplifying the algorithm for finding
the path index when points culled from
the path.

Another item was accounting for empty
paths.

Another change was in the mouse move
callback.  Code was added to pick up
the time index with duplicated time
stamps mixed with culled points.

Last change was due to accounting for
floating point error in timestamps
when unit scaling and scale/biasing
time.  Instead of an equality check,
needed dt < epsilon.