Speedup drag-n-dropping RUNs on multiple plots
A DP file with 180 plots would come up fast,
but dropping a RUN that would add a curve
to each of those 180 plots - koviz would come
to a crawl.

The reason for the slowdown was that koviz was
over-updating the plots.  To fix the issue, instead
of having the CurvesView::dataChanged() handle
pixmap generation, the dataChanged() callback marks the
pixmap as dirty and then when the paintEvent() occurs
the pixmap is updated.  This technique is
the canonical way Qt apps update - dataChanged()
marks the widget as dirty and subsequent painting
handles the pixmap change.

Another issue was the way koviz was synchronizing
across plots.  For example, when panning, all
plots pan together.  The sync logic was in the
dataChanged() callback.  When updating 180 plots,
after each plot update, koviz was updating all
other plots.  This was O(n^2).

This commit takes the sync logic out of the dataChanged()
callback and puts it in the BookModel::setPlotMathRect()
call.  This is another change that goes with
Qt's canonical way - dataChanged() callbacks
should never change the model.