Refresh plots when drag-n-dropping RUN
Since using the pixmap dirty technique to keep from
recreating the plot pixmap, there is now no need to block signals
when a run is added/deleted to/from a set of plots.

This comes after finding the plots would not refresh
all the time when dragging and dropping a run on
the runs list.