Mouse hover not showing point when time duplicated
In one of the test cases, mouse hovering over a curve
to view the point on curve, the live arrow marker wouldn't always
show up.  Turns out, it was an issue with having
multiple duplicated timestamps where some of the duplicated
timestamps were culled with logscale.  The mouse event
needed an update.  The mouse event was not updating the
live time index as it should have.