Better live marker mouse tracking
This commit does two things:

1) It takes the index of the time of the binary search for point,
but then widens the search from that index.

2) With duplicate time points, use the mouse position
to determine time index of path element closest to mouse
instead of maxY.  The maxY wasn't related to
mouse position at all, so didn't really work well because
the user would expect the point closest to mouse to be
selected.