Choose point close to mouse when very close to curve
The mouse to point tracking works well for the most
part - choosing max, min and flats.  This works
well for quickly hopping to a spike.  This is what
the user normally wants. However, there are times
when the user can have a specific point in mind
that is not a max, min or flat.  In that case,
this commit will choose a point close to the mouse
if the mouse if very close, within a character (~16pixels).
So, if the user wants a very specific point,
moving the mouse close to the point will choose it.
If the mouse is far away, the standard behavior of quickly
hopping to spike/flat is used.