The -dp2csv options not working when time specified in DP table
The -dp2csv option takes a DP table and extracts the variables
in the table into a csv file.  Problem was when specifying
time in the dp table variable list.  If the user put time
in the list, the generated csv would add an extra column for
time, but zero it out.  The way to make it work was to remove
the time var from the DP table.  This commit makes it so
the user can have time in the DP table and all works as it should.