Introduce -vars2csv commandline option
The -vars2csv option is used to extract a list
of vars from a RUN of trk files into a csv file.
One can specify units, scale and bias as well.
Example:

koviz RUN_test -vars "ball.state.out.position[0] {ft} scale(-1)
                     bias(100),
                     ball.state.out.velocity[0] {mph}"
               -vars2csv posvel.csv

The way this works underneath assumes the option is
for trk files, so need to work this so any supported
data file/dir can be used for extracting a set of
vars into a csv.