Introduce jpg commandline options for generating images
This introduces the -jpg and -jpgSize commandline options.
If the DP file has more than one pages, koviz will produce
a set of images based on the jpg name e.g.
if DP_a has 3 pages, the following:

    koviz RUN_a DP_a -jpg a.jpg

will produce:

     a.0.jpg
     a.1.jpg
     a.2.jpg

The size of the jpg will default to 2K which is 2560x1440.
If another size is desired, use the -jpgSize option e.g.:

    koviz RUN_a DP_a -jpg a.jpg -jpgSize 3840x2160